dataset/requirements.txt
google-cloud-speech>=1.2.0
google-cloud-storage>=1.19.1
